+++ packages/core/contracts/test/ignore-errors/SafeContractWithLibraryImport.sol
@@ -1,0 +1,8 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.9;
+
+import "./RiskyLibrary.sol";
+
+contract SafeContractWithLibraryImport {
+
+}