+++ packages/core/contracts/test/ignore-errors/UnsafeContractWithTransitiveLibraryCall.sol
@@ -1,0 +1,10 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.9;
+
+import "./TransitiveRiskyLibrary.sol";
+
+contract UnsafeContractWithTransitiveLibraryCall {
+    function unsafe(address target, bytes memory data) public {
+        TransitiveRiskyLibrary.internalDelegateCall(target, data);
+    }
+}